[feat] SwitchDetail에서 사용하는 modal 위치 수정
