+++ src/routes/root/home/SwitchDetailScreen/modals/DeleteItemModal.tsx
